# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tienda Magic API",
        "description": "Documentación de la API de Tienda Magic Swagger OpenApi",
        "contact": {
            "email": "soporte@tiendamagic.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/mod/forums/{forumId}/threads": {
            "get": {
                "tags": [
                    "Moderation"
                ],
                "summary": "Lista de hilos del foro para moderadores",
                "description": "Obtiene la lista de los hilos del foro para moderar.",
                "operationId": "8a90a33c99aacdf4c2bc1aef6e39e923",
                "parameters": [
                    {
                        "name": "forumId",
                        "in": "path",
                        "description": "ID del foro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de hilos de moderación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/threads": {
            "get": {
                "tags": [
                    "Forums"
                ],
                "summary": "Lista global de threads",
                "description": "Obtiene los últimos hilos del foro ordenados por popularidad, fecha, etc.",
                "operationId": "3785412e3613d5d46cdfed8d1d9e3935",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Orden (hot, top, new, commented)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "hot"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de hilos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Forums"
                ],
                "summary": "Crear hilo de foro",
                "description": "Crea un nuevo hilo (thread) vinculado al usuario activo.",
                "operationId": "8ca3cb77a17ec654764eda1282753c9b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "forum_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hilo creado exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/forum/search": {
            "get": {
                "tags": [
                    "Forums"
                ],
                "summary": "Buscar en foros",
                "description": "Busca dentro de los hilos de los foros por título o contenido.",
                "operationId": "d9eb7a09e4d216338df2a9bdef3e0d48",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Término de búsqueda (min 3 caracteres)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de búsqueda"
                    }
                }
            }
        },
        "/api/threads/{thread}": {
            "get": {
                "tags": [
                    "Forums"
                ],
                "summary": "Ver hilo del foro",
                "description": "Obtiene los detalles del hilo y carga sus comentarios raíz.",
                "operationId": "084d35a65f16b8b77ce34642c98c4a42",
                "parameters": [
                    {
                        "name": "thread",
                        "in": "path",
                        "description": "ID del hilo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thread cargado (incrementa views)"
                    }
                }
            },
            "put": {
                "tags": [
                    "Forums"
                ],
                "summary": "Editar hilo",
                "description": "Edita el cuerpo o el título de un hilo (solo para el autor).",
                "operationId": "803e228a1ebd229a9b8ca4234ee30313",
                "parameters": [
                    {
                        "name": "thread",
                        "in": "path",
                        "description": "ID del hilo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hilo editado exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Forums"
                ],
                "summary": "Eliminar hilo",
                "description": "Borra el hilo (soft-delete) si eres el autor.",
                "operationId": "4786e4059ea81e6d2cb3b7f7fc2592e9",
                "parameters": [
                    {
                        "name": "thread",
                        "in": "path",
                        "description": "ID del hilo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensaje correcto"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}